дз 12 выполнено
