+++ src/main/java/ru/tegneryadnov/java/basic/homeworks/homework12/FileInform.java
@@ -1,16 +1,23 @@
 package ru.tegneryadnov.java.basic.homeworks.homework12;
 
-public class FileInform {
-    public FileType fileType;
-    public String fullName;
+import java.io.File;
 
-    public FileInform(String fullName, FileType fileType) {
-        this.fullName = fullName;
-        this.fileType = fileType;
+/**
+ * Класс со списком файлов текущего каталога.
+ * fileType типа FileType используется для сортировки
+ * Папки отображаются в угловых скобках
+ */
+public class FileInform {
+    public final File file;
+    public final FileType fileType;
+
+    public FileInform(File file) {
+        this.file = file;
+        this.fileType = file.isDirectory() ? FileType.DIRECTORY : FileType.FILE;
     }
 
     @Override
     public String toString() {
-        return fileType == FileType.DIRECTORY ? String.format("[%s]", fullName) : fullName;
+        return fileType.equals(FileType.DIRECTORY) ? String.format("<%s>", file.getName()) : file.getName();
     }
 }
