--- src/main/java/ru/tegneryadnov/java/basic/homeworks/homework12_1/FileType.java
@@ -1,6 +1,0 @@
-package ru.tegneryadnov.java.basic.homeworks.homework12_1;
-
-public enum FileType {
-    DIRECTORY,
-    FILE
-}
